XBC-Recommendation-Engine/requirements.txt
azure-functions==1.23.0
Flask==3.1.1
numpy==2.0.2
pandas==2.3.0
requests==2.32.4
azure-storage-blob==12.16.0
scikit-learn==1.6.1
matplotlib==3.9.4
seaborn==0.13.2
azure-identity